+++ app/src/main/java/no/hvl/dat153/thenamequizapp/activities/MainActivity.java
@@ -1,4 +1,4 @@
-package no.hvl.dat153.thenamequizapp;
+package no.hvl.dat153.thenamequizapp.activities;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -7,6 +7,8 @@ import android.os.Bundle;
 import android.util.Log;
 import android.view.View;
 import android.widget.Button;
+
+import no.hvl.dat153.thenamequizapp.R;
 
 public class MainActivity extends AppCompatActivity implements View.OnClickListener {
