+++ app/src/main/java/no/hvl/dat153/thenamequizapp/activities/ResultActivity.java
@@ -1,4 +1,4 @@
-package no.hvl.dat153.thenamequizapp;
+package no.hvl.dat153.thenamequizapp.activities;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -6,6 +6,8 @@ import android.widget.Button;
 import android.widget.TextView;
 
 import androidx.appcompat.app.AppCompatActivity;
+
+import no.hvl.dat153.thenamequizapp.R;
 
 public class ResultActivity extends AppCompatActivity {
 
@@ -26,7 +28,7 @@ public class ResultActivity extends AppCompatActivity {
         textView.setText(display);
 
         final Button returnToStart = findViewById(R.id.returnToStartBtn);
-        Intent MainActivity = new Intent(this, MainActivity.class);
+        Intent MainActivity = new Intent(this, no.hvl.dat153.thenamequizapp.activities.MainActivity.class);
         returnToStart.setOnClickListener(v -> startActivity(MainActivity));
     }
 }
