+++ app/src/main/java/no/hvl/dat153/thenamequizapp/roomdatabase/PersonRepository.java
@@ -1,4 +1,4 @@
-package no.hvl.dat153.thenamequizapp;
+package no.hvl.dat153.thenamequizapp.roomdatabase;
 
 import android.app.Application;
