+++ app/src/main/res/layout/activity_add_entry.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:gravity="center"
     android:orientation="vertical"
-    tools:context=".AddEntryActivity">
+    tools:context=".activities.AddEntryActivity">
 
     <ImageView
         android:id="@+id/imageViewPersonPicture"
